+++ src/routes/run.ts
@@ -133,7 +133,9 @@ export const runCI = async (req: Request, res: Response) => {
   } catch (e) {
     logger.log("CI-JOE failed: ", e);
     await setFailureCommitStatus(commitStatusURL);
+    return res.sendStatus(500);
   }
+  return res.sendStatus(200);
 };
 
 type RepositoryInfo = {
